Pulling in an updated design

+++ src/components/stickyNote.vue
@@ -1,21 +1,29 @@
 <template>
-    <div class="stickyNote" v-if="!this.note.reported">
-        {{ this.note.fingerPrint }}
-        <div v-if="editNoteModalVisible == false" class="stickyNote__content" v-html="computedNote"></div>
-        <button @click="editModal" v-if="showButton()">Edit note</button>
+    <li class="stickyNote" v-if="!this.note.reported">
+        <div class="stickyNote__inner flex flex-column">
+            <div class="stickyNote__header flex flex-justify-right">
+                <button @click="editModal" v-if="showButton()">Edit note</button>
+                <button @click="showConfirmModal(confirmModalType.delete)" v-if="showButton()">Delete note</button>
+                <button @click="showConfirmModal(confirmModalType.report)" v-if="editNoteModalVisible == false">Report</button>
+            </div>
 
-        <div class="stickyNote__edit" v-if="editNoteModalVisible">
-            <button class="stickyNote__edit-close" @click="closeEditNoteModal()">X</button>
-            <textarea v-model="editedNote"></textarea>
-            <button class="stickyNote__edit-done" @click="editNote(editedNote)">Done</button>
-            <span v-if="illegalContent" class="stickyNote__edit-alert">Illegal content, please try again.</span>
+            <div class="stickyNote__content flex-item-1">
+                <template v-if="editNoteModalVisible == false">
+                    <span>{{ this.note.fingerPrint }}</span>
+                    <p v-html="computedNote" class="stickyNote__content-user-content"></p>
+                </template>
+                
+                <div class="stickyNote__edit" v-if="editNoteModalVisible">
+                    <button class="stickyNote__edit-close" @click="closeEditNoteModal()">X</button>
+                    <textarea v-model="editedNote"></textarea>
+                    <button class="stickyNote__edit-done" @click="editNote(editedNote)">Done</button>
+                    <span v-if="illegalContent" class="stickyNote__edit-alert">Illegal content, please try again.</span>
+                </div>
+
+                <confirm-modal v-if="this.confirmModalVisible == true" @confirm="confirm" @deny="deny" :type="selectedModal"></confirm-modal>
+            </div>
         </div>
-
-        <button @click="confirmModalVisible = true" v-if="showButton()">Delete note</button>
-        <confirm-modal v-if="this.confirmModalVisible == true" @confirm="confirm" @deny="deny"></confirm-modal>
-
-        <button @click="reportNote()">Report</button>
-    </div>
+    </li>
 </template>
 
 <script>
@@ -36,7 +44,18 @@
                 confirmModalVisible: false,
                 editNoteModalVisible: false,
                 editedNote: '',
-                illegalContent: false
+                illegalContent: false,
+                confirmModalType: {
+                    delete: {
+                        key: 'delete',
+                        string: 'Are you sure you want to delete your note?  This will delete it forever!'
+                    },
+                    report: {
+                        key: 'report',
+                        string: 'Do you want to report this note?'
+                    },
+                },
+                selectedModal: {}
             }
         },
         props: {
@@ -50,11 +69,23 @@
                     return false
                 }
             },
-            confirm() {
-                this.confirmModalVisible = false;
-                this.deleteStickyNote(this.note);
+            showConfirmModal(type) {
+                this.confirmModalVisible = true;
+                this.selectedModal = type;
             },
-            deny() {this.confirmModalVisible = false;},
+            confirm(type) {
+                this.confirmModalVisible = false;
+
+                if(type.key === 'delete') {
+                    this.deleteStickyNote(this.note);
+                }
+                if(type.key === 'report') {
+                    this.reportStickyNote(this.note);
+                }
+            },
+            deny() {
+                this.confirmModalVisible = false;
+            },
             editModal: function() {
                 if(this.editedNote == '') {
                     this.editedNote = this.note.note;
@@ -84,9 +115,6 @@
                 this.editedNote = '';
                 this.editNoteModalVisible = false;
                 this.illegalContent = false;
-            },
-            reportNote() {
-                this.reportStickyNote(this.note);
             }
         },
         computed: {
@@ -103,8 +131,39 @@
     @import '../styles/variables.less';
 
     .stickyNote {
-        background: @stickyYellow;
-        flex: 1 1 25%;
-        height: 370px;
+        padding: 0 15px 15px 0;
+        position: relative;
+        overflow: hidden;
+        &::before {
+            content: '';
+            display: block;
+            background: @black;
+            height: calc(100% - 15px);
+            width: calc(100% - 15px);
+            position: absolute;
+            top: 15px;
+            left: 15px;
+            border-radius: 24px;
+        }
+        &__inner {
+            overflow: hidden;
+            border: 4px solid @black;
+            border-radius: 24px;
+            width: calc(100% - 8px);
+            height: calc(100% - 8px);
+            position: relative;
+            background: @stickyYellow;
+        }
+
+        &__header {
+            height: 32px;
+            border-bottom: 4px solid @black;
+        }
+        &__content {
+            padding: 24px 24px 80px 24px;
+            &-user-content {
+                margin: 0;
+            }
+        }
     }
 </style>
